requirements.txt
# ps3recomp toolchain dependencies
pycryptodome>=3.19
capstone>=5.0
construct>=2.10
tabulate>=0.9
tomli>=2.0; python_version < "3.11"
tqdm>=4.65
